created MockWallet
