+++ key-wallet-manager/Cargo.toml
@@ -14,6 +14,7 @@ std = ["key-wallet/std", "dashcore/std", "dashcore_hashes/std", "secp256k1/std"]
 serde = ["dep:serde", "key-wallet/serde", "dashcore/serde"]
 getrandom = ["key-wallet/getrandom"]
 bincode = ["dep:bincode","key-wallet/bincode"]
+test-utils = []
 
 [dependencies]
 key-wallet = { path = "../key-wallet", default-features = false }
@@ -24,11 +25,11 @@ serde = { version = "1.0", default-features = false, features = ["derive"], opti
 async-trait = "0.1"
 bincode = { version = "=2.0.0-rc.3", optional = true }
 zeroize = { version = "1.8", features = ["derive"] }
+tokio = { version = "1.32", default-features = false }
 
 [dev-dependencies]
 hex = "0.4"
 serde_json = "1.0"
-tokio = { version = "1.32", features = ["full"] }
 
 [lints.rust]
 unexpected_cfgs = { level = "allow", check-cfg = ['cfg(bench)', 'cfg(fuzzing)'] }
