+++ key-wallet-manager/src/test_utils/mod.rs
@@ -1,0 +1,3 @@
+mod wallet;
+
+pub use wallet::MockWallet;
